Разделил тесты для каждой политики замещения.
