+++ tests/test_cache_decorator.py
@@ -20,32 +20,6 @@ def test_caching():
     assert slow_function.call_count == 1  # Функция должна быть вызвана только один раз
 
 
-def test_cache_eviction_LRU():
-    """Проверяет, что кэш очищает старые значения при превышении глубины  при политике LRU."""
-
-    @cache(depth=2)
-    def slow_function(x):
-        slow_function.call_count += 1
-        return x * x
-
-    # Инициализируем счетчик вызовов
-    slow_function.call_count = 0
-
-    slow_function(1)
-    slow_function(2)
-    assert slow_function.call_count == 2  # Должно быть 2 вызова
-
-    slow_function(2)
-    assert slow_function.call_count == 2
-    # Вызов с новым значением вытесняет менее используемое значение (1)
-    slow_function(3)
-    assert slow_function.call_count == 3  # Еще один вызов
-
-    # Повторный вызов для 1 должен пересчитаться, так как он вытеснен
-    slow_function(1)
-    assert slow_function.call_count == 4  # Пересчитывается заново
-
-
 def test_multiple_functions():
     """Проверяет, что кэш работает независимо для нескольких функций."""
 
@@ -95,83 +69,3 @@ def test_cache_with_kwargs():
     # Новый вызов с другим значением ключевого аргумента
     assert slow_function(3, power=3) == 27
     assert slow_function.call_count == 2  # Должен быть вызов функции
-
-
-def test_cache_eviction_FIFO():
-    """Проверяет, что кэш очищает старые значения при превышении глубины при политике FIFO."""
-
-    @cache(depth=2, policy="FIFO")
-    def slow_function(x):
-        slow_function.call_count += 1
-        return x * x
-
-    slow_function.call_count = 0
-
-    slow_function(1)
-    slow_function(2)
-    assert slow_function.call_count == 2
-
-    slow_function(2)
-    assert slow_function.call_count == 2
-
-    slow_function(3)
-    assert slow_function.call_count == 3
-
-    slow_function(1)
-    assert slow_function.call_count == 4
-
-    slow_function(2)
-    assert slow_function.call_count == 5
-
-
-def test_cache_eviction_LIFO():
-    """Проверяет, что кэш очищает старые значения при превышении глубины при политике LIFO."""
-
-    @cache(depth=2, policy="LIFO")
-    def slow_function(x):
-        slow_function.call_count += 1
-        return x * x
-
-    # Инициализируем счетчик вызовов
-    slow_function.call_count = 0
-
-    slow_function(1)
-    slow_function(2)
-    assert slow_function.call_count == 2
-
-    slow_function(2)
-    assert slow_function.call_count == 2
-
-    slow_function(3)
-    assert slow_function.call_count == 3
-
-    slow_function(1)
-    assert slow_function.call_count == 3
-
-    slow_function(3)
-    assert slow_function.call_count == 3
-
-
-def test_cache_eviction_MRU():
-    """Проверяет, что кэш очищает старые значения при превышении глубины при политике LIFO."""
-
-    @cache(depth=2, policy="MRU")
-    def slow_function(x):
-        slow_function.call_count += 1
-        return x * x
-
-    # Инициализируем счетчик вызовов
-    slow_function.call_count = 0
-
-    slow_function(1)
-    slow_function(2)
-    assert slow_function.call_count == 2
-
-    slow_function(2)
-    assert slow_function.call_count == 2
-
-    slow_function(3)
-    assert slow_function.call_count == 3
-
-    slow_function(1)
-    assert slow_function.call_count == 3
